Refactor: move DurabilityLikeInfo out
